@@ -2,7 +2,7 @@
   "name": "path-slashes",
   "version": "1.1.1",
   "description": "Adding or removing of trailing and/or leading slashes in strings, test for their existence or join strings with slashes without duplicating them.",
-  "author": "Simon Lepel <simbo@simbo.de> (https://simbo.de)",
+  "author": "Simon Lepel <simbo@simbo.de> (https://simbo.de/)",
   "license": "MIT",
   "keywords": [
     "path",
@@ -13,61 +13,46 @@
     "trailing"
   ],
   "homepage": "https://github.com/simbo/path-slashes#readme",
-  "repository": "github:simbo/path-slashes",
-  "bugs": "https://github.com/simbo/path-slashes/issues",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/simbo/path-slashes.git"
+  },
+  "bugs": {
+    "url": "https://github.com/simbo/path-slashes/issues"
+  },
   "engines": {
     "node": ">=16"
   },
-  "source": "src/path-slashes.ts",
-  "main": "dist/path-slashes.js",
-  "umd:main": "dist/path-slashes.umd.js",
-  "module": "dist/path-slashes.m.js",
-  "esmodule": "dist/path-slashes.modern.js",
-  "types": "dist/path-slashes.d.ts",
+  "type": "module",
+  "main": "dist/index.js",
   "files": [
     "dist/",
-    "LICENSE",
+    "LICENSE.md",
     "README.md"
   ],
   "scripts": {
-    "build": "microbundle",
-    "build:watch": "microbundle watch",
-    "clean": "rm -rf ./dist",
-    "clean:coverage": "rm -rf ./coverage",
-    "coverage:open": "open ./coverage/lcov-report/index.html",
-    "lint": "eslint -c .eslintrc --ext .ts,.js .",
-    "lint:fix": "eslint -c .eslintrc --ext .ts,.js --fix .",
+    "build": "tsc --project ./tsconfig.dist.json",
+    "lint": "eslint -c .eslintrc.json .",
+    "lint:fix": "eslint -c .eslintrc.json --fix .",
     "format": "prettier --check .",
     "format:fix": "prettier --write .",
-    "prebuild": "npm run clean",
-    "prebuild:watch": "npm run clean",
     "preflight": "npm run lint && npm run format && npm run test && npm run build",
-    "pretest": "npm run clean:coverage",
-    "test": "jest",
-    "test:watch": "npm run test --watch",
-    "upgrade": "npx npm-check-updates -u -t minor && npx npm-check-updates --interactive --reject typescript,@types/node && npm i"
+    "test": "NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest",
+    "test:watch": "npm run test -- --watch",
+    "upgrade": "npx --yes npm-check-updates -u -t minor && npx --yes npm-check-updates --interactive --reject typescript,@types/node && npm i"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-typescript": "^7.12.7",
     "@ianvs/prettier-plugin-sort-imports": "^3.7.2",
-    "@types/jest": "^26.0.20",
+    "@types/jest": "^29.5.1",
     "@types/node": "^16.18.25",
-    "@typescript-eslint/eslint-plugin": "^4.13.0",
-    "@typescript-eslint/parser": "^4.13.0",
-    "babel-jest": "^26.6.3",
-    "eslint": "^7.18.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-editorconfig": "^3.0.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsdoc": "^31.0.5",
-    "eslint-plugin-prefer-arrow": "^1.2.2",
-    "eslint-plugin-unicorn": "^26.0.1",
-    "jest": "^26.6.3",
-    "microbundle": "^0.13.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-unicorn": "^46.0.0",
+    "jest": "^29.5.0",
     "prettier": "^2.8.8",
-    "ts-jest": "^26.4.4",
-    "typescript": "^4.1.3"
+    "ts-jest": "^29.1.0",
+    "typescript": "^4.9.5"
   }
 }
